requirements.txt
git+git://github.com/Rapptz/discord.py
jishaku
discord-py-slash-command